requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
functions-framework>=3.0.0
